requirements.txt
# Core MkDocs
mkdocs>=1.5.3
mkdocs-material>=9.5.3

# PDF Processing
PyMuPDF>=1.23.8
Pillow>=10.1.0

# Additional utilities
pyyaml>=6.0.1
jinja2>=3.1.2

# Development
pytest>=7.4.3
black>=23.12.1
flake8>=6.1.0
